+++ core.py
@@ -26,820 +26,8 @@ import pathlib # Added for path manipulation
 from datetime import datetime # Added for debug timestamps
 from bpy.types import Operator
 from bpy.props import StringProperty, EnumProperty, BoolProperty # Added BoolProperty
-from . import preferences # Import the preferences module
-from .preferences import BM_Preferences, ITEM_DEFINITIONS # Import specific names
-
-
-def prefs():
-    """
-    Directly retrieves the addon's preferences.
-    Assumes bpy.context and addon preferences are always accessible.
-    """
-    user_preferences = bpy.context.preferences
-    return user_preferences.addons[__package__].preferences
-
-def find_versions(filepath):
-    version_list = []
-    _start_time_fv = None
-    if prefs().debug:
-        _start_time_fv = datetime.now()
-        print(f"DEBUG: find_versions START for path: {filepath}")
-
-    if not filepath or not os.path.isdir(filepath):
-        if prefs().debug:
-            print(f"DEBUG: find_versions: filepath invalid or not a directory: {filepath}")
-        return version_list
-
-    try:
-        _listdir_start_time_fv = None
-        if prefs().debug:
-            _listdir_start_time_fv = datetime.now()
-            print(f"DEBUG: find_versions CALLING os.listdir for path: {filepath}")
-        for file_or_dir_name in os.listdir(filepath):
-            path = os.path.join(filepath, file_or_dir_name)
-            if os.path.isdir(path):      
-                version_list.append((file_or_dir_name, file_or_dir_name, ''))
-        if prefs().debug:
-            _listdir_end_time_fv = datetime.now()
-            print(f"DEBUG: (took: {(_listdir_end_time_fv - _listdir_start_time_fv).total_seconds():.6f}s) find_versions FINISHED os.listdir for path: {filepath}")
-    except OSError as e: # Catch specific OS errors like PermissionError
-        if prefs().debug:
-            print(f"DEBUG: find_versions: Error accessing filepath {filepath}: {e}")
-    
-    if prefs().debug and _start_time_fv:
-        # print("\nVersion List: ", version_list)
-        _end_time_fv = datetime.now()
-        # Consider summarizing 'version_list' if it can be very long.
-        # For now, printing the full list for detailed debugging.
-        print(f"DEBUG: (took: {(_end_time_fv - _start_time_fv).total_seconds():.6f}s) find_versions END for path: '{filepath}', found {len(version_list)} versions. List: {version_list}")
-
-    return version_list
-
-# --- UIList definitions for Backup/Restore Item Configuration ---
-# ITEM_DEFINITIONS and BM_BackupItem (PropertyGroup) are now defined in preferences.py
-# BM_UL_BackupItemsList (UIList) remains here as it's part of the core UI window.
-
-class BM_UL_BackupItemsList(bpy.types.UIList):
-    """UIList for displaying backup/restore items with enabled/shared toggles."""
-    # This UIList will operate on items of type preferences.BM_BackupItem
-    # Revised layout: enabled_check (10%) | item_name (80%) | shared_check (10%)
-    def draw_item(self, context, layout, data, item, icon, active_data, active_propname, index):
-        prefs_for_list = active_data # This will be the BM_Preferences instance
-        prop_enabled_name = f"{prefs_for_list.tabs.lower()}_{item.identifier}" # Use prefs_for_list.tabs
-
-        # Determine icon based on property state
-        is_enabled = getattr(prefs_for_list, prop_enabled_name, False)
-        icon_enabled = 'CHECKBOX_HLT' if is_enabled else 'CHECKBOX_DEHLT'
-
-        is_shared = getattr(prefs_for_list, f"shared_{item.identifier}", False)
-        icon_shared = 'LINKED' if is_shared else 'UNLINKED'
-
-        # Create a main row for this item that will pack its children tightly.
-        # 'layout' is the UILayout object for the entire item's row provided by UIList.
-        item_row = layout.row(align=True)
-
-        # Enabled icon (leftmost)
-        item_row.prop(prefs_for_list, prop_enabled_name, text="", icon=icon_enabled, icon_only=True, emboss=False)
-
-        # Item Name (middle, will expand next to the enabled icon)
-        item_row.label(text=item.name)
-
-        # Shared icon (rightmost)
-        # This sub-row will be pushed to the right by the expanding label.
-        shared_icon_container = item_row.row(align=True)
-        shared_icon_container.alignment = 'RIGHT' # Align the icon itself to the right of this container
-        shared_icon_container.prop(prefs_for_list, f"shared_{item.identifier}", text="", icon=icon_shared, icon_only=True, emboss=False)
-
-
-class OT_AbortOperation(Operator):
-    """Operator to signal cancellation of the ongoing backup/restore operation."""
-    bl_idname = "bm.abort_operation"
-    bl_label = "Abort Backup/Restore"
-    bl_description = "Requests cancellation of the current operation"
-
-    def execute(self, context):
-        prefs().abort_operation_requested = True
-        # The modal operator will pick this up and handle the actual cancellation.
-        if prefs().debug:
-            print("DEBUG: OT_AbortOperation executed, abort_operation_requested set to True.")
-        return {'FINISHED'}
-
-class OT_QuitBlenderNoSave(bpy.types.Operator):
-    """Quits Blender without saving current user preferences."""
-    bl_idname = "bm.quit_blender_no_save"
-    bl_label = "Quit & Restart Blender"
-    bl_description = "Attempts to start a new Blender instance, then quits the current one. If 'Save on Quit' is enabled in Blender's preferences, you will be warned."
-
-    @classmethod
-    def poll(cls, context):
-        return True # Always allow attempting to quit
-
-    def invoke(self, context, event):
-        # It's good practice to get addon_prefs once, especially if debug is checked multiple times.
-        # However, ensure prefs() doesn't fail if context is minimal during early invoke.
-        addon_prefs_instance = None
-        try:
-            addon_prefs_instance = prefs()
-        except Exception as e:
-            print(f"DEBUG: OT_QuitBlenderNoSave.invoke: Could not retrieve addon preferences for debug logging: {e}")
-        
-        prefs_main = context.preferences # Use bpy.context.preferences for global auto-save setting
-
-        if addon_prefs_instance and addon_prefs_instance.debug:
-            print(f"DEBUG: OT_QuitBlenderNoSave.invoke(): Context preferences type: {type(context.preferences)}, "
-                  f"Has 'use_preferences_save': {hasattr(context.preferences, 'use_preferences_save')}")
-
-        # The original logic, now with the debug prints above it.
-        # This line will still raise an AttributeError if 'use_preferences_save' is missing.
-        # The debug output should help understand why.
-        if prefs_main and hasattr(prefs_main, 'use_preferences_save') and prefs_main.use_preferences_save:
-            # 'Save on Quit' is ON. We need to warn the user.
-            return context.window_manager.invoke_confirm(self, event)
-        else:
-            # 'Save on Quit' is OFF, or attribute is missing (hasattr was False), or prefs_view is None.
-            # If attribute is missing, we proceed as if it's OFF to avoid the dialog,
-            # but the execute method will log this uncertainty.
-            return self.execute(context)
-    
-    def execute(self, context):
-        prefs_main = context.preferences # Use bpy.context.preferences
-        addon_prefs = prefs() # Get addon preferences for debug
-        
-        if addon_prefs.debug:
-            print(f"DEBUG: OT_QuitBlenderNoSave.execute():")
-            print(f"  context: {context}")
-            print(f"  context.preferences: {context.preferences}")
-            print(f"  type(prefs_main): {type(prefs_main)}")
-            if prefs_main:
-                try:
-                    print(f"  dir(prefs_main): {dir(prefs_main)}")
-                except Exception as e_dir:
-                    print(f"  Error getting dir(prefs_main): {e_dir}")
-                print(f"  Has 'use_preferences_save' attribute?: {hasattr(prefs_main, 'use_preferences_save')}")
-
-        # This is the line from the traceback.
-        # We check hasattr again to be safe and for clearer logging.
-        blender_will_save_on_quit = False # Default assumption
-        if prefs_main and hasattr(prefs_main, 'use_preferences_save'):
-            blender_will_save_on_quit = prefs_main.use_preferences_save
-        elif prefs_main: # prefs_main exists but hasattr was False
-            if addon_prefs.debug:
-                print("WARNING: OT_QuitBlenderNoSave.execute: 'use_preferences_save' attribute missing on Preferences object. Assuming Blender will save preferences for safety.")
-            blender_will_save_on_quit = True # Assume worst-case for logging if attribute is missing
-        else: # prefs_view is None
-             if addon_prefs.debug:
-                print("WARNING: OT_QuitBlenderNoSave.execute: prefs_view is None. Assuming Blender will save preferences for safety.")
-             blender_will_save_on_quit = True
-
-        # --- Attempt to launch new Blender instance ---
-        blender_exe = bpy.app.binary_path
-        new_instance_launched_successfully = False
-        if blender_exe and blender_exe is not None and os.path.exists(blender_exe): # Check if path is valid and not None
-            try:
-                if addon_prefs.debug:
-                    print(f"DEBUG: OT_QuitBlenderNoSave: Attempting to launch new Blender instance from: {blender_exe}")
-
-                args = [blender_exe]
-                kwargs = {}
-
-                if sys.platform == "win32":
-                    DETACHED_PROCESS = 0x00000008 # subprocess.DETACHED_PROCESS
-                    kwargs['creationflags'] = DETACHED_PROCESS
-                elif sys.platform == "darwin": # macOS
-                    pass # No special flags usually needed
-                elif sys.platform.startswith("linux"): # Linux and other POSIX
-                    kwargs['start_new_session'] = True
-
-                try:
-                    subprocess.Popen(args, **kwargs)
-                except FileNotFoundError as e:
-                    if addon_prefs.debug:
-                        print(f"ERROR: OT_QuitBlenderNoSave: Blender executable not found: {e}")
-                        import traceback
-                        traceback.print_exc()
-                except PermissionError as e:
-                    if addon_prefs.debug:
-                        print(f"ERROR: OT_QuitBlenderNoSave: Permission denied when launching Blender: {e}")
-                        import traceback
-                        traceback.print_exc()
-                except Exception as e:
-                    if addon_prefs.debug:
-                        print(f"ERROR: OT_QuitBlenderNoSave: Unexpected error occurred: {e}")
-                        import traceback
-                        traceback.print_exc()
-
-                new_instance_launched_successfully = True
-                if addon_prefs.debug:
-                    print(f"DEBUG: OT_QuitBlenderNoSave: New Blender instance launch command issued.")
-                    if sys.platform == "win32" and 'creationflags' in kwargs:
-                         print(f"DEBUG: OT_QuitBlenderNoSave: Using creationflags={kwargs['creationflags']}")
-                    elif 'start_new_session' in kwargs:
-                        print(f"DEBUG: OT_QuitBlenderNoSave: Using start_new_session={kwargs['start_new_session']}")
-
-            except Exception as e:
-                if addon_prefs.debug:
-                    print(f"ERROR: OT_QuitBlenderNoSave: Failed to launch new Blender instance: {e}")
-        else:
-            if addon_prefs.debug:
-                print(f"DEBUG: OT_QuitBlenderNoSave: Blender executable path not found or invalid: '{blender_exe}'. Skipping new instance launch.")
-
-        if blender_will_save_on_quit:
-            # This path is taken if invoke_confirm was accepted by the user.
-            if addon_prefs.debug:
-                print("DEBUG: OT_QuitBlenderNoSave: Quitting Blender. 'Save on Quit' is ON. "
-                      "User was warned; preferences WILL be saved by Blender.")
-            # The warning should have made it clear that preferences *will* be saved.
-        else:
-            # 'Save on Quit' is OFF.
-            if addon_prefs.debug: print("DEBUG: OT_QuitBlenderNoSave: Quitting Blender. 'Save on Quit' is OFF. Preferences will NOT be saved by Blender.")
-        
-        if addon_prefs.debug:
-            if new_instance_launched_successfully:
-                print("DEBUG: OT_QuitBlenderNoSave: New instance launch command succeeded. Proceeding to quit current instance.")
-            else:
-                print("DEBUG: OT_QuitBlenderNoSave: New instance launch failed or was skipped. Proceeding to quit current instance.")
-
-        # --- Proceed with quitting the current Blender instance ---
-        bpy.ops.wm.quit_blender()
-        return {'FINISHED'}
-
-    # This draw method is for the confirmation dialog when use_save_on_quit is True
-    def draw(self, context):
-        layout = self.layout
-        col = layout.column()
-        col.label(text="Blender's 'Save Preferences on Quit' is currently ENABLED.", icon='ERROR')
-        col.separator()
-        col.label(text="If you proceed, Blender will save its current in-memory preferences when quitting.")
-        col.label(text="This will likely overwrite the 'userpref.blend' file you just restored.")
-        col.separator()
-        col.label(text="To ensure the restored 'userpref.blend' is used on next startup:")
-        box = col.box()
-        box.label(text="1. Select 'Cancel' on this dialog (see details below).")
-        box.label(text="2. Go to: Edit > Preferences > Save & Load.")
-        box.label(text="3. Uncheck the 'Save on Quit' option.")
-        box.label(text="4. Manually quit Blender (File > Quit).")
-        box.label(text="5. Restart Blender. Your restored preferences should now be active.")
-        box.label(text="   (You can re-enable 'Save on Quit' after restarting, if desired).")
-        col.separator()
-        col.label(text=f"Choosing '{self.bl_label}' (OK) below will quit this Blender session,")
-        col.label(text="and it WILL save its current preferences due to the global setting.")
-        col.label(text="An attempt will then be made to start a new Blender instance.")
-        col.separator()
-        col.label(text="Choosing 'Cancel' will abort this quit/restart attempt by the addon.")
-
-
-class OT_CloseReportDialog(bpy.types.Operator):
-    """Closes the report dialog without taking further action."""
-    bl_idname = "bm.close_report_dialog"
-    bl_label = "Don't Quit Now"
-    bl_options = {'INTERNAL'}
-
-    def execute(self, context):
-        # This operator's purpose is just to be a clickable item in the popup
-        # that allows the popup to close without triggering the quit sequence.
-        if prefs().debug:
-            print("DEBUG: OT_CloseReportDialog executed (User chose not to quit/restart from report).")
-        return {'FINISHED'}
-
-
-class OT_ShowFinalReport(bpy.types.Operator):
-    """Operator to display a popup message. Used by timers for deferred reports."""
-    bl_idname = "bm.show_final_report"
-    bl_label = "Show Operation Report"
-    bl_options = {'INTERNAL'} # This operator is not meant to be called directly by the user from UI search
-
-    # Static class variables to hold the report data
-    _title: str = "Report"
-    _icon: str = "INFO"
-    _lines: list = []
-    _timer = None # Timer for the modal part
-    _show_restart_button: bool = False
-    _restart_operator_idname: str = ""
-
-    @classmethod
-    def set_report_data(cls, lines, title, icon, show_restart=False, restart_op_idname=""):
-        """Sets the data to be displayed by the popup."""
-        cls._lines = lines
-        cls._title = title
-        cls._icon = icon
-        cls._show_restart_button = show_restart
-        cls._restart_operator_idname = restart_op_idname
-        if prefs().debug:
-            print(f"DEBUG: OT_ShowFinalReport.set_report_data: Title='{cls._title}', Icon='{cls._icon}', Lines={len(cls._lines)}, ShowRestart={cls._show_restart_button}, RestartOp='{cls._restart_operator_idname}'")
-
-    def execute(self, context): #@NoSelf
-        """Displays the popup report directly."""
-        if prefs().debug:
-            print(f"DEBUG: OT_ShowFinalReport.execute: Displaying popup. Title='{OT_ShowFinalReport._title}'")
-
-        # --- This function defines what's drawn in the popup ---
-        def draw_for_popup(self_menu, context_inner): # 'self_menu' is the Menu instance, 'context_inner' is the context for the popup
-            layout = self_menu.layout
-            for line in OT_ShowFinalReport._lines:
-                layout.label(text=line)
-            
-            if OT_ShowFinalReport._show_restart_button and OT_ShowFinalReport._restart_operator_idname:
-                layout.separator()
-                # Place the restart and close buttons in a single row for side-by-side layout
-                row = layout.row(align=True)
-                row.operator(OT_ShowFinalReport._restart_operator_idname, icon='FILE_REFRESH')
-                row.operator(OT_CloseReportDialog.bl_idname, icon='CANCEL')
-
-        context.window_manager.popup_menu(draw_for_popup, title=OT_ShowFinalReport._title, icon=OT_ShowFinalReport._icon)
-        return {'FINISHED'}
-
-
-class OT_BackupManagerWindow(Operator):
-    """Open the Backup Manager window."""
-    bl_idname = "bm.open_backup_manager_window"
-    bl_label = "Backup Manager"
-    bl_options = {'REGISTER'} # No UNDO needed for a UI window    
-
-    _cancelled: bool = False # Instance variable to track cancellation state
-
-    def _update_window_tabs(self, context):
-        """Ensures BM_Preferences.tabs is updated when this window's tabs change, triggering searches."""
-        # --- Early exit conditions for stale instance ---
-        # Use getattr for robustness, in case _cancelled is not yet set on self during an early update call
-        if getattr(self, '_cancelled', False):
-            # Minimal logging if possible, avoid complex operations
-            # print(f"DEBUG: OT_BackupManagerWindow._update_window_tabs - Bailing out: _cancelled is True. Self: {self}")
-            return
-
-        prefs_instance = None
-        try:
-            prefs_instance = prefs()
-            if not prefs_instance: # If prefs() returns None or an invalid object
-                # print(f"DEBUG: OT_BackupManagerWindow._update_window_tabs - Bailing out: prefs() returned None/invalid. Self: {self}")
-                return
-        except Exception as e:
-            # If prefs() itself raises an exception (e.g., addon not found during unregister/register)
-            print(f"ERROR: Backup Manager: Error in OT_BackupManagerWindow._update_window_tabs (accessing prefs): {e}. Self: {self}")
-            return
-        
-        # --- Proceed with original logic if checks passed ---
-        if prefs_instance.tabs != self.tabs:
-            if prefs_instance.debug: # Safe to use prefs_instance.debug now
-                print(f"DEBUG: OT_BackupManagerWindow._update_window_tabs: self.tabs ('{self.tabs}') != prefs.tabs ('{prefs_instance.tabs}'). Updating prefs.tabs. Self: {self}")
-            try:
-                prefs_instance.tabs = self.tabs # This will call BM_Preferences.update_version_list
-            except Exception as e_update:
-                 # Catch errors during the actual update of prefs_instance.tabs
-                 if prefs_instance.debug:
-                    print(f"ERROR: Backup Manager: Error in OT_BackupManagerWindow._update_window_tabs (updating prefs.tabs): {e_update}. Self: {self}")
-                 # Decide if further action is needed, e.g., report error to user or log
-
-
-    tabs: EnumProperty(
-        name="Mode",
-        items=[
-            ("BACKUP", "Backup", "Switch to Backup mode", "COLORSET_03_VEC", 0),
-            ("RESTORE", "Restore", "Switch to Restore mode", "COLORSET_04_VEC", 1)
-        ],
-        default="BACKUP",
-        update=_update_window_tabs
-    )
-
-    show_item_configuration: BoolProperty(
-        name="Configure Items for Selected Operation",
-        description="Expand to configure which items are included in the backup/restore operation",
-        default=False # Start collapsed by default
-    )
-
-    def _draw_path_age(self, layout, path_to_check):
-        """Helper to draw cached path age."""
-        prefs_instance = prefs() # Get prefs for debug flag
-        if not path_to_check or not os.path.isdir(path_to_check): # Basic check before cache lookup
-            layout.label(text="Last change: Path N/A")
-            return
-        display_text = BM_Preferences._age_cache.get(path_to_check)
-        if display_text is None:
-            display_text = "Last change: Calculating..."
-            if prefs_instance.debug:
-                print(f"DEBUG: OT_BackupManagerWindow._draw_path_age: No cache for '{path_to_check}', displaying 'Calculating...'")
-        elif prefs_instance.debug:
-            print(f"DEBUG: OT_BackupManagerWindow._draw_path_age: Using cached value for '{path_to_check}': {display_text}")
-        layout.label(text=display_text)
-
-    def _draw_path_size(self, layout, path_to_check):
-        """Helper to draw cached path size."""
-        prefs_instance = prefs() # Get prefs for debug flag
-        if not path_to_check or not os.path.isdir(path_to_check): # Basic check
-            layout.label(text="Size: Path N/A")
-            return
-        display_text = BM_Preferences._size_cache.get(path_to_check)
-        if display_text is None:
-            display_text = "Size: Calculating..."
-            if prefs_instance.debug:
-                print(f"DEBUG: OT_BackupManagerWindow._draw_path_size: No cache for '{path_to_check}', displaying 'Calculating...'")
-        elif prefs_instance.debug:
-            print(f"DEBUG: OT_BackupManagerWindow._draw_path_size: Using cached value for '{path_to_check}': {display_text}")
-        layout.label(text=display_text)
-
-    def _draw_backup_tab(self, layout, context, prefs_instance):
-        """Draws the Backup tab content."""
-        row_main  = layout.row(align=True) # Main row for From/To/Actions
-        
-        box_from = row_main.box() 
-        col_from = box_from.column()
-
-        if not prefs_instance.advanced_mode:
-            path_from_val = prefs_instance.blender_user_path
-            col_from.label(text = "Backup From: " + str(prefs_instance.active_blender_version), icon='COLORSET_03_VEC')   
-            col_from.label(text = path_from_val)      
-            if prefs_instance.show_path_details:
-                self._draw_path_age(col_from, path_from_val) 
-                self._draw_path_size(col_from, path_from_val)
-
-            box_to = row_main.box() # Add box_to to row_main
-            col_to = box_to.column()
-            path_to_val =  os.path.join(prefs_instance.backup_path, prefs_instance.system_id, str(prefs_instance.active_blender_version)) if prefs_instance.backup_path and prefs_instance.system_id else "N/A"
-            col_to.label(text = "Backup To: " + str(prefs_instance.active_blender_version), icon='COLORSET_04_VEC')
-            col_to.label(text = path_to_val)          
-            if prefs_instance.show_path_details:
-                self._draw_path_age(col_to, path_to_val)    
-                self._draw_path_size(col_to, path_to_val)
-        else: # Advanced mode
-            # --- Backup From Box ---
-            source_version_selected = prefs_instance.backup_versions # This is the string value of the selected item
-            path_from_val = os.path.join(os.path.dirname(prefs_instance.blender_user_path), source_version_selected) if prefs_instance.blender_user_path and source_version_selected else "N/A"
-            
-            col_from.label(text="Backup From: " + source_version_selected, icon='COLORSET_03_VEC')
-            col_from.label(text=path_from_val)
-            if prefs_instance.show_path_details: self._draw_path_age(col_from, path_from_val)
-            if prefs_instance.show_path_details: self._draw_path_size(col_from, path_from_val)
-            col_from.prop(prefs_instance, 'backup_versions', text='Version' if prefs_instance.expand_version_selection else '', expand=prefs_instance.expand_version_selection)
-
-            # --- Backup To Box ---
-            box_to = row_main.box() # Add box_to to row_main
-            col_to = box_to.column()
-            if prefs_instance.custom_version_toggle:
-                target_version_displayed = prefs_instance.custom_version
-                path_to_val = os.path.join(prefs_instance.backup_path, prefs_instance.system_id, target_version_displayed) if prefs_instance.backup_path and prefs_instance.system_id and target_version_displayed else "N/A"
-                col_to.label(text="Backup To: " + target_version_displayed, icon='COLORSET_04_VEC')
-                col_to.label(text=path_to_val)
-                if prefs_instance.show_path_details: self._draw_path_age(col_to, path_to_val)
-                if prefs_instance.show_path_details: self._draw_path_size(col_to, path_to_val)
-                col_to.prop(prefs_instance, 'custom_version', text='Version')
-            else: # Not custom_version_toggle, use restore_versions for dropdown
-                target_version_displayed = prefs_instance.restore_versions # This is the string value of the selected item
-                path_to_val = os.path.join(prefs_instance.backup_path, prefs_instance.system_id, target_version_displayed) if prefs_instance.backup_path and prefs_instance.system_id and target_version_displayed else "N/A"
-                col_to.label(text="Backup To: " + target_version_displayed, icon='COLORSET_04_VEC')
-                col_to.label(text=path_to_val)
-                if prefs_instance.show_path_details: self._draw_path_age(col_to, path_to_val)
-                if prefs_instance.show_path_details: self._draw_path_size(col_to, path_to_val)
-                col_to.prop(prefs_instance, 'restore_versions', text='Version' if prefs_instance.expand_version_selection else '', expand=prefs_instance.expand_version_selection)
-
-        # --- Actions Column ---
-        col_actions = row_main.column() # For main action buttons and global toggles
-        col_actions.scale_x = 0.9 # Slightly narrower for this column
-        col_actions.operator("bm.run_backup_manager", text="Backup Selected", icon='COLORSET_03_VEC').button_input = 'BACKUP' 
-        if prefs_instance.advanced_mode:
-            col_actions.operator("bm.run_backup_manager", text="Backup All", icon='COLORSET_03_VEC').button_input = 'BATCH_BACKUP' 
-        col_actions.separator(factor=1.0)
-        col_actions.prop(prefs_instance, 'dry_run')  
-        col_actions.prop(prefs_instance, 'clean_path')  
-        col_actions.prop(prefs_instance, 'advanced_mode') 
-        if prefs_instance.advanced_mode:
-            col_actions.prop(prefs_instance, 'custom_version_toggle')  
-            col_actions.prop(prefs_instance, 'expand_version_selection')    
-            col_actions.separator(factor=1.0)
-            col_actions.operator("bm.run_backup_manager", text="Delete Backup", icon='COLORSET_01_VEC').button_input = 'DELETE_BACKUP'
-
-
-    def _draw_restore_tab(self, layout, context, prefs_instance):
-        """Draws the Restore tab content."""
-        row_main  = layout.row(align=True) # Main row for From/To/Actions
-
-        box_from = row_main.box()
-        col_from = box_from.column()
-
-        if not prefs_instance.advanced_mode:
-            path_from_val = os.path.join(prefs_instance.backup_path, prefs_instance.system_id, str(prefs_instance.active_blender_version)) if prefs_instance.backup_path and prefs_instance.system_id else "N/A"
-            col_from.label(text = "Restore From: " + str(prefs_instance.active_blender_version), icon='COLORSET_04_VEC')   
-            col_from.label(text = path_from_val)                  
-            if prefs_instance.show_path_details: self._draw_path_age(col_from, path_from_val); self._draw_path_size(col_from, path_from_val)
-            
-            box_to = row_main.box() # Add box_to to row_main
-            col_to = box_to.column()
-            path_to_val =  prefs_instance.blender_user_path
-            col_to.label(text = "Restore To: " + str(prefs_instance.active_blender_version), icon='COLORSET_03_VEC')   
-            col_to.label(text = path_to_val)              
-            if prefs_instance.show_path_details: self._draw_path_age(col_to, path_to_val); self._draw_path_size(col_to, path_to_val)
-        else: # Advanced Mode
-            source_ver = prefs_instance.restore_versions
-            path_from_val = os.path.join(prefs_instance.backup_path, prefs_instance.system_id, source_ver) if prefs_instance.backup_path and prefs_instance.system_id and source_ver else "N/A"
-            col_from.label(text="Restore From: " + source_ver, icon='COLORSET_04_VEC')
-            col_from.label(text=path_from_val)
-            if prefs_instance.show_path_details: self._draw_path_age(col_from, path_from_val); self._draw_path_size(col_from, path_from_val)
-            col_from.prop(prefs_instance, 'restore_versions', text='Version' if prefs_instance.expand_version_selection else '', expand=prefs_instance.expand_version_selection)
-
-            box_to = row_main.box() # Add box_to to row_main
-            col_to = box_to.column()
-            target_ver = prefs_instance.backup_versions
-            path_to_val = os.path.join(os.path.dirname(prefs_instance.blender_user_path), target_ver) if prefs_instance.blender_user_path and target_ver else "N/A"
-            col_to.label(text="Restore To: " + target_ver, icon='COLORSET_03_VEC')
-            col_to.label(text=path_to_val)
-            if prefs_instance.show_path_details: self._draw_path_age(col_to, path_to_val); self._draw_path_size(col_to, path_to_val)
-            col_to.prop(prefs_instance, 'backup_versions', text='Version' if prefs_instance.expand_version_selection else '', expand=prefs_instance.expand_version_selection)
-
-        # --- Actions Column ---
-        col_actions = row_main.column()
-        col_actions.scale_x = 0.9
-        col_actions.operator("bm.run_backup_manager", text="Restore Selected", icon='COLORSET_04_VEC').button_input = 'RESTORE'
-        if prefs_instance.advanced_mode:
-            col_actions.operator("bm.run_backup_manager", text="Restore All", icon='COLORSET_04_VEC').button_input = 'BATCH_RESTORE'
-        col_actions.separator(factor=1.0)
-        col_actions.prop(prefs_instance, 'dry_run')      
-        col_actions.prop(prefs_instance, 'clean_path')   
-        col_actions.prop(prefs_instance, 'advanced_mode')  
-        if prefs_instance.advanced_mode:
-            col_actions.prop(prefs_instance, 'expand_version_selection')
-
-
-    def draw(self, context): # Standard signature for invoke_props_dialog
-        layout = self.layout # Provided by invoke_props_dialog
-        prefs_instance = None
-        _debug_draw = False # Default
-
-        # --- Early exit conditions for stale instance ---
-        if self._cancelled: # Check if the instance itself was marked as cancelled
-            layout.label(text="Window closing (operator cancelled)...")
-            # print(f"DEBUG: OT_BackupManagerWindow.draw() - Bailing out: _cancelled is True. Self: {self}")
-            return
-
-        # Attempt to get preferences and set debug flag safely
-        try:
-            prefs_instance = prefs() # Get current addon preferences
-            if not prefs_instance: # If prefs() returns None or an invalid object
-                layout.label(text="Window closing (preferences unavailable)...")
-                # print(f"DEBUG: OT_BackupManagerWindow.draw() - Bailing out: prefs() returned None/invalid. Self: {self}")
-                return
-            _debug_draw = prefs_instance.debug # Safe to access .debug now
-        except Exception as e:
-            # If prefs() itself raises an exception (e.g., addon not found during unregister/register)
-            layout.label(text=f"Window closing (error accessing preferences: {e})...")
-            # print(f"DEBUG: OT_BackupManagerWindow.draw() - Bailing out: Exception in prefs(): {e}. Self: {self}")
-            return
-
-        # --- Proceed with normal drawing if all checks passed ---
-        try:
-            _start_time_draw_obj = None # Use a different name to avoid conflict if prefs_instance is None initially
-
-            if _debug_draw:
-                # Debug output for draw() start
-                timestamp = datetime.now().strftime("%H:%M:%S.%f")[:-3]
-                progress_val_str = f"{prefs_instance.operation_progress_value:.1f}%" if prefs_instance.show_operation_progress else "N/A (hidden)"
-                op_message = prefs_instance.operation_progress_message if prefs_instance.show_operation_progress else "N/A (hidden)"
-                print(f"DEBUG: [{timestamp}] OT_BackupManagerWindow.draw() CALLED. Progress: {progress_val_str}, Msg: '{op_message}', show_op_progress: {prefs_instance.show_operation_progress}, Tabs: {self.tabs}")
-                _start_time_draw_obj = datetime.now()
-           
-            is_operation_running = prefs_instance.show_operation_progress
-
-
-            # --- Top section for global settings ---
-            box_top = layout.box()
-            col_top = box_top.column(align=True)            
-            col_top.use_property_split = True
-            col_top.separator()
-           
-            # System ID is always read-only display
-            row_system_id = col_top.row()
-            row_system_id.enabled = False
-            row_system_id.prop(prefs_instance, "system_id")            
-
-            # Group settings that should be disabled during an operation
-            settings_to_disable_group = col_top.column()
-            settings_to_disable_group.enabled = not is_operation_running
-            settings_to_disable_group.prop(prefs_instance, 'backup_path')
-            #col_top.prop(prefs_instance, 'ignore_files') # Still commented out
-            settings_to_disable_group.prop(prefs_instance, 'show_path_details')
-            
-            # Debug 
-            col_top.prop(prefs_instance, 'debug')  
-            col_top.separator()   
-
-            if prefs_instance.debug: # Only show system paths if debug is enabled
-                col_top.separator()
-
-                # Blender System Paths (Read-Only)
-                box_system_paths_display = col_top.box()
-                col_system_paths_display = box_system_paths_display.column(align=True)
-                col_system_paths_display.label(text="Blender System Paths (Read-Only - Debug):")
-
-                # Blender Installation Path
-                blender_install_path = os.path.dirname(bpy.app.binary_path)
-                row_install = col_system_paths_display.row(align=True)
-                row_install.label(text="Installation Path:")
-                op_install = row_install.operator(preferences.OT_OpenPathInExplorer.bl_idname, icon='FILEBROWSER', text="")
-                op_install.path_to_open = blender_install_path
-                col_system_paths_display.label(text=blender_install_path)
-
-                # User Version Folder Path
-                row_user_version_folder = col_system_paths_display.row(align=True)
-                row_user_version_folder.label(text="User Version Folder:")
-                op_user_version_folder = row_user_version_folder.operator(preferences.OT_OpenPathInExplorer.bl_idname, icon='FILEBROWSER', text="")
-                op_user_version_folder.path_to_open = prefs_instance.blender_user_path
-                col_system_paths_display.label(text=prefs_instance.blender_user_path)
-
-                # User Config Subfolder Path is already available via prefs_instance.config_path
-                # No need to construct it manually here as it's already a preference property.
-                # We can add it here if desired, similar to the preferences panel.
-
-
-
-            # --- Save Preferences Button ---
-            # Conditionally show the "Save Preferences" button.
-            # Only show it if Blender's "Save Preferences on Quit" is OFF.
-            prefs_main = context.preferences # Use bpy.context.preferences
-            show_manual_save_button = True # Default to showing the button
-
-            if prefs_main and hasattr(prefs_main, 'use_preferences_save'):
-                if prefs_main.use_preferences_save: # If 'Auto-Save Preferences' is ON
-                    show_manual_save_button = False # Then DO NOT show the manual save button
-                    if _debug_draw:
-                        print("DEBUG: OT_BackupManagerWindow.draw(): Blender's 'Save on Quit' is ON. Hiding manual 'Save Preferences' button.")
-                elif _debug_draw: # 'Save on Quit' is OFF
-                     print("DEBUG: OT_BackupManagerWindow.draw(): Blender's 'Save on Quit' is OFF. Showing manual 'Save Preferences' button.")
-            elif _debug_draw: # Could not determine 'Save on Quit' state
-                print("DEBUG: OT_BackupManagerWindow.draw(): Could not determine Blender's 'Save on Quit' state. Defaulting to show manual 'Save Preferences' button.")
-
-            if show_manual_save_button:
-                save_prefs_row = layout.row(align=True)
-                save_prefs_row.enabled = not is_operation_running # Disable if operation is running
-
-                # Check if preferences have unsaved changes (shows '*' in Blender UI)
-                label_text = "Save Preferences"
-                if bpy.context.preferences.is_dirty:
-                    label_text += " *"
-                
-                # Proper alignment for the button within the row
-                save_prefs_row.label(text="") # Spacer on the left
-                save_prefs_button_col = save_prefs_row.column()
-                save_prefs_button_col.scale_x = 0.5 # Make button narrower
-                save_prefs_button_col.operator("wm.save_userpref", text=label_text, icon='PREFERENCES')
-            
-            # --- Tabs for Backup/Restore ---      
-            layout.use_property_split = False
-            layout.prop(self, "tabs", expand=False) # Use the operator's own tabs property
-
-            # --- Progress UI ---
-            if prefs_instance.show_operation_progress:
-                op_status_box = layout.box()
-                op_status_col = op_status_box.column(align=True)
-
-                # Display the progress message
-                if prefs_instance.operation_progress_message:
-                    op_status_col.label(text=prefs_instance.operation_progress_message)
-                
-                # Create a new row for the progress bar and abort button
-                progress_row = op_status_col.row(align=True)
-
-                # Display the progress value as a slider (without its own text label)
-                # operation_progress_value is a 0.0-100.0 factor. The text="" hides the label to its left.
-                progress_row.prop(prefs_instance, "operation_progress_value", slider=True, text="")
-                                
-                # Abort button
-                progress_row.operator("bm.abort_operation", text="", icon='CANCEL') # Text removed for compactness
-
-            # --- Tab content for Backup/Restore --- 
-            tab_content_box = layout.box() # Box for the content of the selected tab
-            tab_content_box.enabled = not is_operation_running # Disable tab content if operation is running
-
-            if self.tabs == "BACKUP":
-                self._draw_backup_tab(tab_content_box, context, prefs_instance)
-            elif self.tabs == "RESTORE":
-                self._draw_restore_tab(tab_content_box, context, prefs_instance)
-            
-            # --- Item Selection List (common to both tabs, drawn after tab-specific content) ---
-            # This list is now part of the main window, drawn below Backup/Restore specific controls.
-            # It's always visible when advanced_mode is on, regardless of the active tab.
-            if prefs_instance.advanced_mode:
-                layout.separator() # Add some space before the list
-                # Use a box for the collapsible section
-                item_config_box = layout.box()
-                # Draw the collapsible header using the operator's property
-                item_config_box.prop(self, "show_item_configuration", icon="TRIA_DOWN" if self.show_item_configuration else "TRIA_RIGHT", icon_only=False, emboss=False)
-
-                if self.show_item_configuration:
-                    # Ensure the collection in preferences is populated
-                    prefs_instance._ensure_backup_items_populated() # Call method on prefs_instance
-
-                    item_config_box.template_list(
-                        "BM_UL_BackupItemsList", "", # UIList class name (now defined in core.py)
-                        prefs_instance, "backup_items_collection",  # data_ptr (prefs), prop_name (in prefs)
-                        prefs_instance, "active_backup_item_index", # active_data_ptr (prefs), active_propname (in prefs)
-                        # ITEM_DEFINITIONS is now in preferences module
-                        rows=len(preferences.ITEM_DEFINITIONS) if len(preferences.ITEM_DEFINITIONS) <= 10 else 10)
-
-            if _debug_draw and _start_time_draw_obj:
-                _end_time_draw_obj = datetime.now()
-                print(f"DEBUG: (took: {(_end_time_draw_obj - _start_time_draw_obj).total_seconds():.6f}s) OT_BackupManagerWindow.draw() END")
-                print("-" * 70) # Add a separator line
-
-        except Exception as e:
-            print(f"ERROR: Backup Manager: Error during OT_BackupManagerWindow.draw() main block: {e}")
-            layout.label(text=f"Error drawing Backup Manager window: {e}")
-
-
-    def execute(self, context):
-        prefs_instance = None
-        try:
-            prefs_instance = prefs()
-        except Exception as e:
-            # Log error and attempt to cancel gracefully
-            print(f"ERROR: OT_BackupManagerWindow.execute() - Failed to get preferences: {e}. Self: {self}")
-            self.cancel(context) # Ensure timer cleanup
-            return {'CANCELLED'} # Indicate failure
-
-        if prefs_instance.debug:
-            print(f"DEBUG: OT_BackupManagerWindow.execute() ENTER. Context: {context}, Self: {self}")
-        
-        # Ensure timer is cleaned up if execute is called (e.g. by an "OK" button if one existed)
-        if self._timer:
-            context.window_manager.event_timer_remove(self._timer)
-            self._timer = None
-            if prefs_instance.debug: print(f"DEBUG: OT_BackupManagerWindow.execute() - Timer removed.")
-        
-        # Check prefs_instance again in case it was None from the try-except block
-        # but cancel didn't lead to an immediate return (though it should).
-        # This is more for defensive logging.
-        if prefs_instance and prefs_instance.debug:
-            print(f"DEBUG: OT_BackupManagerWindow.execute() EXIT. Returning {{'FINISHED'}}. Self: {self}")
-        return {'FINISHED'} # Signal successful completion.
-
-    def invoke(self, context, event):
-        self._cancelled = False # Reset cancellation flag on new invocation
-
-        prefs_instance = prefs()
-
-        # Critical check: If preferences are not available, cancel immediately.
-        if not prefs_instance:
-            print(f"ERROR: OT_BackupManagerWindow.invoke() - Failed to get addon preferences. Cannot initialize window. Self: {self}")
-            return {'CANCELLED'}
-
-        # Now it's safe to use prefs_instance
-        _debug_active = prefs_instance.debug # Store debug state for local use
-
-        if _debug_active:
-            print(f"DEBUG: OT_BackupManagerWindow.invoke() CALLED. Initializing tabs from prefs: {prefs_instance.tabs}")
-        
-        self.tabs = prefs_instance.tabs # Initialize window tabs from preferences
-
-        # Explicitly trigger the version list update (which includes the SEARCH operator call)
-        # and subsequent path detail scan (if show_path_details is true).
-        # This ensures that the necessary scans are performed on first window open,
-        # as the update chain via self.tabs -> _update_window_tabs -> prefs.tabs
-        # might not fire if the tab values are already synchronized.
-        if _debug_active:
-            print(f"DEBUG: OT_BackupManagerWindow.invoke() - Explicitly calling prefs_instance.update_version_list(context) for initial scan.")
-        prefs_instance._update_backup_path_and_versions(context)
-
-        # Use invoke_props_dialog to open the window.
-        # The operator's draw() method will be called by Blender to populate the dialog.
-        result = context.window_manager.invoke_props_dialog(self, width=700)
-        if _debug_active:
-            print(f"DEBUG: OT_BackupManagerWindow.invoke() EXIT. invoke_props_dialog returned: {result}. Self: {self}")
-        # If invoke_props_dialog returns {'RUNNING_MODAL'}, our modal() method will also run.
-        return result
-
-    # The modal() method is removed as the operator is no longer self-modal.
-    # invoke_props_dialog handles the dialog's modality.
-
-    def cancel(self, context):
-        self._cancelled = True # Mark this instance as cancelled
-
-        # Use the robust prefs() function from core.py
-        _debug_active = False # Default to False for safety during cancel
-        prefs_instance_for_cancel = None
-        try:
-            prefs_instance_for_cancel = prefs()
-            if prefs_instance_for_cancel:
-                _debug_active = prefs_instance_for_cancel.debug
-        except Exception:
-            pass # Ignore errors getting prefs during cancel, prioritize cleanup
-        if _debug_active: print(f"DEBUG: OT_BackupManagerWindow.cancel() ENTER. Context: {context}, Self: {self}, _cancelled set to True.")
-        
-        # If an operation (from OT_BackupManager) is in progress,
-        # DO NOT request it to abort just because this UI window is closing.
-        # The OT_BackupManager is its own modal operator and can be cancelled
-        # via its own ESC handling or the explicit Abort button (if window was open).
-        try:
-            if prefs_instance_for_cancel and prefs_instance_for_cancel.show_operation_progress: # Check if OT_BackupManager is likely active
-                if _debug_active:
-                    print(f"DEBUG: OT_BackupManagerWindow.cancel() - Operation is in progress (show_operation_progress is True).")
-                    print(f"DEBUG: OT_BackupManagerWindow.cancel() - Window is closing, but the background operation will NOT be aborted by this window's cancellation.")
-                # Explicitly DO NOT set: prefs_instance_for_cancel.abort_operation_requested = True
-            elif prefs_instance_for_cancel: # No operation in progress
-                 if _debug_active:
-                    print(f"DEBUG: OT_BackupManagerWindow.cancel() - No operation in progress (show_operation_progress is False). Window closing normally.")
-        except Exception as e:
-            if _debug_active: print(f"DEBUG: OT_BackupManagerWindow.cancel() - Error accessing prefs_instance_for_cancel.show_operation_progress: {e}. Self: {self}")
-            
-        if _debug_active:
-            print(f"DEBUG: OT_BackupManagerWindow.cancel() EXIT. Self: {self}")
-        # Operator.cancel() is expected to do cleanup.
-        # If invoke_props_dialog calls this, it handles the {'CANCELLED'} state internally.
+from . import preferences # For BM_Preferences, ITEM_DEFINITIONS
+from . import utils # For get_addon_preferences, find_versions # ui will be imported locally
                 
 class OT_BackupManager(Operator):
     ''' run backup & restore '''
@@ -871,57 +59,57 @@ class OT_BackupManager(Operator):
     def create_ignore_pattern(self):
         self.ignore_backup.clear()
         self.ignore_restore.clear()
-        list = [x for x in regular_expression.split(',|\s+', prefs().ignore_files) if x!='']        
+        list = [x for x in regular_expression.split(',|\s+', utils.get_addon_preferences().ignore_files) if x!='']        
         for item in list:
             self.ignore_backup.append(item)
             self.ignore_restore.append(item)
 
-        if not prefs().backup_bookmarks:
+        if not utils.get_addon_preferences().backup_bookmarks:
             self.ignore_backup.append('bookmarks.txt')
-        if not prefs().restore_bookmarks:
-            self.ignore_restore.append('bookmarks.txt')
-        if not prefs().backup_recentfiles:
+        if not utils.get_addon_preferences().restore_bookmarks:
+            self.ignore_restore.append('bookmarks.txt') # Use utils.get_addon_preferences()
+        if not utils.get_addon_preferences().backup_recentfiles:
             self.ignore_backup.append('recent-files.txt')
-        if not prefs().restore_recentfiles:
-            self.ignore_restore.append('recent-files.txt')   
+        if not utils.get_addon_preferences().restore_recentfiles:
+            self.ignore_restore.append('recent-files.txt') # Use utils.get_addon_preferences()
 
-        if not prefs().backup_startup_blend:
+        if not utils.get_addon_preferences().backup_startup_blend:
             self.ignore_backup.append('startup.blend')
-        if not prefs().restore_startup_blend:
-            self.ignore_restore.append('startup.blend')            
-        if not prefs().backup_userpref_blend:
+        if not utils.get_addon_preferences().restore_startup_blend:
+            self.ignore_restore.append('startup.blend') # Use utils.get_addon_preferences()
+        if not utils.get_addon_preferences().backup_userpref_blend:
             self.ignore_backup.append('userpref.blend')
-        if not prefs().restore_userpref_blend:
-            self.ignore_restore.append('userpref.blend')            
-        if not prefs().backup_workspaces_blend:
+        if not utils.get_addon_preferences().restore_userpref_blend:
+            self.ignore_restore.append('userpref.blend') # Use utils.get_addon_preferences()
+        if not utils.get_addon_preferences().backup_workspaces_blend:
             self.ignore_backup.append('workspaces.blend')
-        if not prefs().restore_workspaces_blend:
-            self.ignore_restore.append('workspaces.blend')  
+        if not utils.get_addon_preferences().restore_workspaces_blend:
+            self.ignore_restore.append('workspaces.blend') # Use utils.get_addon_preferences()
 
-        if not prefs().backup_cache:
+        if not utils.get_addon_preferences().backup_cache:
             self.ignore_backup.append('cache')
-        if not prefs().restore_cache:
-            self.ignore_restore.append('cache')
+        if not utils.get_addon_preferences().restore_cache:
+            self.ignore_restore.append('cache') # Use utils.get_addon_preferences()
 
-        if not prefs().backup_datafile:
+        if not utils.get_addon_preferences().backup_datafile:
             self.ignore_backup.append('datafiles')
-        if not prefs().restore_datafile:
-            self.ignore_restore.append('datafiles')
+        if not utils.get_addon_preferences().restore_datafile:
+            self.ignore_restore.append('datafiles') # Use utils.get_addon_preferences()
 
-        if not prefs().backup_addons:
+        if not utils.get_addon_preferences().backup_addons:
             self.ignore_backup.append('addons')
-        if not prefs().restore_addons:
-            self.ignore_restore.append('addons')
+        if not utils.get_addon_preferences().restore_addons:
+            self.ignore_restore.append('addons') # Use utils.get_addon_preferences()
             
-        if not prefs().backup_extensions:
+        if not utils.get_addon_preferences().backup_extensions:
             self.ignore_backup.append('extensions')
-        if not prefs().restore_extensions:
+        if not utils.get_addon_preferences().restore_extensions:
             self.ignore_restore.append('extensions')
 
-        if not prefs().backup_presets:
+        if not utils.get_addon_preferences().backup_presets:
             self.ignore_backup.append('presets')
-        if not prefs().restore_presets:
-            self.ignore_restore.append('presets')
+        if not utils.get_addon_preferences().restore_presets:
+            self.ignore_restore.append('presets') # Use utils.get_addon_preferences()
     
     def cancel(self, context):
         """Ensures timer and progress UI are cleaned up if the operator is cancelled externally."""
@@ -929,7 +117,7 @@ class OT_BackupManager(Operator):
         _debug_active = False # Default to False for safety during cancel
         prefs_instance_for_cancel = None
         try:
-            prefs_instance_for_cancel = prefs()
+            prefs_instance_for_cancel = utils.get_addon_preferences()
             if prefs_instance_for_cancel:
                 _debug_active = prefs_instance_for_cancel.debug
         except Exception:
@@ -971,14 +159,15 @@ class OT_BackupManager(Operator):
 
     @staticmethod
     def _deferred_show_report_static(message_lines, title, icon, show_restart=False, restart_op_idname=""):
-        if prefs().debug: 
+        from . import ui # Import ui locally
+        if utils.get_addon_preferences().debug: 
             print(f"DEBUG: _deferred_show_report_static: Preparing to invoke bm.show_final_report. Title='{title}', ShowRestart={show_restart}, RestartOp='{restart_op_idname}'")
-        OT_ShowFinalReport.set_report_data(lines=message_lines, 
+        ui.OT_ShowFinalReport.set_report_data(lines=message_lines, 
                                            title=title, 
                                            icon=icon, 
                                            show_restart=show_restart, 
                                            restart_op_idname=restart_op_idname)
-        bpy.ops.bm.show_final_report('EXEC_DEFAULT') # Changed from INVOKE_SCREEN
+        bpy.ops.bm.show_final_report('EXEC_DEFAULT')
         return None # Stop the timer
 
     # Keep the instance method for direct calls if needed, though static is preferred for deferred.
@@ -994,8 +183,8 @@ class OT_BackupManager(Operator):
         is_shared = False
         if item_identifier_or_None:
             # Check against actual ITEM_DEFINITIONS identifiers
-            # ITEM_DEFINITIONS is imported from preferences module
-            if item_identifier_or_None in [item[0] for item in ITEM_DEFINITIONS]:
+            # ITEM_DEFINITIONS is available via preferences.ITEM_DEFINITIONS
+            if item_identifier_or_None in [item[0] for item in preferences.ITEM_DEFINITIONS]:
                  is_shared = getattr(prefs_instance, f"shared_{item_identifier_or_None}", False)
 
         if self.current_operation_type == 'BACKUP':
@@ -1021,7 +210,7 @@ class OT_BackupManager(Operator):
             self.report({'WARNING'}, f"Source path does not exist or is not a directory: {self.current_source_path}")
             return False
 
-        prefs_instance = prefs()
+        prefs_instance = utils.get_addon_preferences()
         self.create_ignore_pattern()
         current_ignore_list = self.ignore_backup if self.current_operation_type == 'BACKUP' else self.ignore_restore
 
@@ -1048,7 +237,7 @@ class OT_BackupManager(Operator):
         # For RESTORE, this specific variable isn't directly used in this function as dest is always local.
         # The _prepare_restore_files_from_source handles its own version_name logic for iterating source.
 
-        if prefs().debug:
+        if prefs_instance.debug:
             print(f"Preparing file list for {self.current_operation_type}")
             print(f"Source: {self.current_source_path}")
             print(f"Target: {self.current_target_path}")
@@ -1080,15 +269,15 @@ class OT_BackupManager(Operator):
                 dest_file = ""
                 if self.current_operation_type == 'BACKUP':
                     # Determine the item_identifier for this file to check its shared_status
-                    effective_item_id_for_shared_check = None
-                    if filename in [item[0] for item in ITEM_DEFINITIONS]: # ITEM_DEFINITIONS from preferences
+                    effective_item_id_for_shared_check = None # Use preferences.ITEM_DEFINITIONS
+                    if filename in [item[0] for item in preferences.ITEM_DEFINITIONS]:
                         effective_item_id_for_shared_check = filename
                     else:
                         if relative_dir_part and relative_dir_part != '.':
-                            # Check parts of the relative path (e.g., "addons" in "scripts/addons")
+                            # Check parts of the relative path
                             parts = pathlib.Path(relative_dir_part).parts
                             for part in reversed(parts): # Check deeper parts first
-                                if part in [item[0] for item in ITEM_DEFINITIONS]:
+                                if part in [item[0] for item in preferences.ITEM_DEFINITIONS]:
                                     effective_item_id_for_shared_check = part
                                     break
                     
@@ -1106,7 +295,7 @@ class OT_BackupManager(Operator):
                     dest_file = os.path.join(self.current_target_path, path_segment_in_version)
 
                 if os.path.normpath(src_file) == os.path.normpath(dest_file):
-                    if prefs().debug: print(f"Skipping copy, source and destination are the same file: {src_file}")
+                    if prefs_instance.debug: print(f"Skipping copy, source and destination are the same file: {src_file}")
                     continue
 
                 # Ensure no duplicates if this function is somehow called in a way that could overlap
@@ -1117,7 +306,7 @@ class OT_BackupManager(Operator):
                     print(f"DEBUG: _prepare_file_list: Duplicate file pair skipped: ({src_file}, {dest_file})")
         
         self.total_files = len(self.files_to_process)
-        if prefs().debug:
+        if prefs_instance.debug:
             print(f"Total files to process: {self.total_files}")
         return True
 
@@ -1133,7 +322,7 @@ class OT_BackupManager(Operator):
         Returns {'RUNNING_MODAL'} if a new item is started modally,
         {'FINISHED'} if batch is complete or no items to process initially.
         """
-        pref_instance = prefs() # Get fresh preferences
+        pref_instance = utils.get_addon_preferences() # Get fresh preferences
 
         if self.current_batch_item_index < self.total_batch_items:
             source_path, target_path, op_type, version_name = self.batch_operations_list[self.current_batch_item_index]
@@ -1221,6 +410,7 @@ class OT_BackupManager(Operator):
             if self.batch_operations_list:
                  overall_op_type = self.batch_operations_list[0][2] # Get op_type from first item
                  report_title = f"Batch {overall_op_type.capitalize()} Report"
+                 from . import ui # Import for OT_QuitBlenderNoSave.bl_label
 
             show_restart_btn_batch = False
             if overall_op_type == 'RESTORE': # Show restart info even on dry run for simulation
@@ -1234,7 +424,7 @@ class OT_BackupManager(Operator):
                 lambda final_report_lines=final_report_lines, report_title=report_title, show_restart_btn_batch=show_restart_btn_batch:
                     OT_BackupManager._deferred_show_report_static(
                         final_report_lines, report_title, 'INFO', show_restart=show_restart_btn_batch, restart_op_idname="bm.quit_blender_no_save"
-                    ),
+                    ), # This will call ui.OT_ShowFinalReport via _deferred_show_report_static
                 first_interval=0.01
                 )
 
@@ -1249,7 +439,7 @@ class OT_BackupManager(Operator):
             return {'FINISHED'}
 
     def modal(self, context, event):
-        pref_instance = prefs() # Get fresh preferences
+        pref_instance = utils.get_addon_preferences() # Get fresh preferences
         # Capture the state of the abort request flag at the beginning of this modal event
         was_aborted_by_ui_button = pref_instance.abort_operation_requested
         
@@ -1277,7 +467,10 @@ class OT_BackupManager(Operator):
                 
                 cancel_message = f"{op_description} cancelled by user."
                 self.report({'WARNING'}, cancel_message)
-                bpy.app.timers.register(lambda: OT_BackupManager._deferred_show_report_static([cancel_message], "Operation Cancelled", "WARNING"), first_interval=0.01)
+                bpy.app.timers.register(
+                    lambda: OT_BackupManager._deferred_show_report_static([cancel_message], "Operation Cancelled", "WARNING"), 
+                    first_interval=0.01
+                ) # This will call ui.OT_ShowFinalReport
                 
                 pref_instance.operation_progress_message = cancel_message
                 pref_instance.operation_progress_value = 0.0 
@@ -1331,6 +524,7 @@ class OT_BackupManager(Operator):
                     if self.current_operation_type == 'RESTORE': # Show restart info even on dry run for simulation
                         report_message_lines.append("") # Add a blank line for spacing
                         report_message_lines.append("IMPORTANT: For restored settings to fully apply, this Blender session must be ended.")
+                        from . import ui # Import for OT_QuitBlenderNoSave.bl_label
                         report_message_lines.append(f"Use the '{OT_QuitBlenderNoSave.bl_label}' button below.")
                         show_restart_btn = True
 
@@ -1342,8 +536,8 @@ class OT_BackupManager(Operator):
                     op_type_for_report_title = self.current_operation_type
                     
                     bpy.app.timers.register(lambda: OT_BackupManager._deferred_show_report_static(
-                        report_message_lines, f"{op_type_for_report_title} Report", report_icon,
-                        show_restart=show_restart_btn, restart_op_idname="bm.quit_blender_no_save"
+                        report_message_lines, f"{op_type_for_report_title} Report", report_icon, # This will call ui.OT_ShowFinalReport
+                        show_restart=show_restart_btn, restart_op_idname=ui.OT_QuitBlenderNoSave.bl_idname # ui already imported if show_restart_btn is True
                     ), first_interval=0.01)
                     self.report({'INFO'}, " ".join(report_message_lines)) 
                     
@@ -1373,13 +567,13 @@ class OT_BackupManager(Operator):
                 return {'PASS_THROUGH'} # Let the next cycle handle termination via top conditions
 
             # Process a batch of files
-            for _ in range(BM_Preferences.FILES_PER_TICK_MODAL_OP): # Use constant from preferences
+            for _ in range(preferences.BM_Preferences.FILES_PER_TICK_MODAL_OP): # Use constant from preferences
                 if not self.files_to_process:
                     break # No more files in the list for this tick
 
                 src_file, dest_file = self.files_to_process.pop(0)
 
-                if not prefs().dry_run:
+                if not pref_instance.dry_run:
                     try:
                         os.makedirs(os.path.dirname(dest_file), exist_ok=True)
                         shutil.copy2(src_file, dest_file)
@@ -1430,12 +624,12 @@ class OT_BackupManager(Operator):
         ultimate_target_dir: The local Blender version path (e.g., .../Blender/4.1/)
         process_shared_state: True if we are looking for items that are marked as shared in prefs.
         """
-        prefs_instance = prefs()
+        prefs_instance = utils.get_addon_preferences()
         # self.ignore_restore should already be set by create_ignore_pattern() in execute()
         current_ignore_list = self.ignore_restore
 
         shared_item_identifiers_globally = { # All items marked as shared in prefs
-            item_def[0] for item_def in ITEM_DEFINITIONS # ITEM_DEFINITIONS from preferences
+            item_def[0] for item_def in preferences.ITEM_DEFINITIONS # Use preferences.ITEM_DEFINITIONS
             if getattr(prefs_instance, f"shared_{item_def[0]}", False)
         }
 
@@ -1483,10 +677,9 @@ class OT_BackupManager(Operator):
         return {'PASS_THROUGH'} # Allow other events to be processed
 
     def execute(self, context): 
-        pref_instance = prefs()
-        pref_backup_versions = BM_Preferences.backup_version_list
-        pref_restore_versions = BM_Preferences.restore_version_list
-        pref_restore_versions = BM_Preferences.restore_version_list
+        pref_instance = utils.get_addon_preferences()
+        pref_backup_versions = preferences.BM_Preferences.backup_version_list
+        pref_restore_versions = preferences.BM_Preferences.restore_version_list
 
         if pref_instance.debug:
             print("\n\nbutton_input: ", self.button_input)                    
@@ -1543,7 +736,7 @@ class OT_BackupManager(Operator):
                         version_name_for_operation = pref_instance.restore_versions # This is the version name from backup list
 
                 if self.button_input == 'BACKUP':
-                    if prefs().clean_path:
+                    if pref_instance.clean_path:
                         # Determine version name for cleaning backup paths
                         version_to_clean_name = ""
                         if not pref_instance.advanced_mode:
@@ -1559,16 +752,16 @@ class OT_BackupManager(Operator):
                         default_target_to_clean = os.path.join(pref_instance.backup_path, pref_instance.system_id, version_to_clean_name)
 
                         if os.path.exists(default_target_to_clean):
-                            if prefs().debug: print(f"Attempting to clean default backup path: {default_target_to_clean}")
+                            if pref_instance.debug: print(f"Attempting to clean default backup path: {default_target_to_clean}")
                             try:
-                                if not prefs().dry_run: shutil.rmtree(default_target_to_clean)
+                                if not pref_instance.dry_run: shutil.rmtree(default_target_to_clean)
                             except OSError as e: self.report({'WARNING'}, f"Failed to clean {default_target_to_clean}: {e}")
                         # Clean shared target path
                         shared_target_to_clean = os.path.join(pref_instance.backup_path, OT_BackupManager.SHARED_FOLDER_NAME, version_to_clean_name)
                         if os.path.exists(shared_target_to_clean):
-                            if prefs().debug: print(f"Attempting to clean shared backup path: {shared_target_to_clean}")
+                            if pref_instance.debug: print(f"Attempting to clean shared backup path: {shared_target_to_clean}")
                             try:
-                                if not prefs().dry_run: shutil.rmtree(shared_target_to_clean)
+                                if not utils.get_addon_preferences().dry_run: shutil.rmtree(shared_target_to_clean)
                             except OSError as e: self.report({'WARNING'}, f"Failed to clean {shared_target_to_clean}: {e}")
                     if not self._prepare_file_list(): return {'CANCELLED'} # Populates self.files_to_process for BACKUP
 
@@ -1610,7 +803,8 @@ class OT_BackupManager(Operator):
 
                     bpy.app.timers.register(lambda: OT_BackupManager._deferred_show_report_static(
                         _msg_lines, _title, _icon
-                    ), first_interval=0.01)
+                    ), first_interval=0.01) # This will call ui.OT_ShowFinalReport
+
                     return {'FINISHED'}
 
                 initial_message = f"Starting {self.current_operation_type}... ({self.total_files} files)"
@@ -1683,9 +877,9 @@ class OT_BackupManager(Operator):
 
                 if os.path.exists(target_path):
                     try:
-                        if not prefs().dry_run:
+                        if not pref_instance.dry_run:
                             shutil.rmtree(target_path)
-                        action_verb = "Would delete" if prefs().dry_run else "Deleted"
+                        action_verb = "Would delete" if pref_instance.dry_run else "Deleted"
                         report_msg_line1 = f"{action_verb} backup:"
                         report_msg_line2 = target_path
                         final_report_lines = [report_msg_line1, report_msg_line2]
@@ -1694,13 +888,14 @@ class OT_BackupManager(Operator):
                         version_name_of_deleted = os.path.basename(target_path)
                         shared_path_to_delete = os.path.join(pref_instance.backup_path, OT_BackupManager.SHARED_FOLDER_NAME, version_name_of_deleted)
                         if os.path.exists(shared_path_to_delete):
-                            if not prefs().dry_run:
+                            if not pref_instance.dry_run:
                                 shutil.rmtree(shared_path_to_delete)
                             final_report_lines.append(f"{action_verb} shared backup part:")
                             final_report_lines.append(shared_path_to_delete)
                             if pref_instance.debug or pref_instance.dry_run: print(f"{action_verb} Shared Backup Part: {shared_path_to_delete}")
                         
                         self.report({'INFO'}, " ".join(final_report_lines))
+                        # This will call ui.OT_ShowFinalReport
                         bpy.app.timers.register(
                             lambda lines=final_report_lines: OT_BackupManager._deferred_show_report_static(lines, "Delete Backup Report", 'COLORSET_01_VEC'),
                             first_interval=0.01
@@ -1713,23 +908,24 @@ class OT_BackupManager(Operator):
                         error_msg_line1 = f"{action_verb} {target_path}:"
                         error_msg_line2 = str(e)
                         self.report({'WARNING'}, f"{error_msg_line1} {error_msg_line2}")
+                        # This will call ui.OT_ShowFinalReport
                         bpy.app.timers.register(
                             lambda lines=[error_msg_line1, error_msg_line2]: OT_BackupManager._deferred_show_report_static(lines, "Delete Backup Error", 'WARNING'),
                             first_interval=0.01
                         )
-                        # OT_BackupManager.ShowReport_static(message=[error_msg_line1, error_msg_line2], title="Delete Backup Error", icon='WARNING')
-                        if prefs().debug: # Keep print for debug
+                        if utils.get_addon_preferences().debug: # Keep print for debug
                             print(f"\n{action_verb} {target_path}: {e}")
                 else:
                     not_found_msg = f"Not found, nothing to delete: {target_path}"
                     self.report({'INFO'}, not_found_msg)
+                    # This will call ui.OT_ShowFinalReport
                     bpy.app.timers.register(lambda: OT_BackupManager._deferred_show_report_static([f"Not found, nothing to delete:", target_path], "Delete Backup Report", 'INFO'), first_interval=0.01)
                     if pref_instance.debug: # Keep print for debug
                         print(f"\nBackup to delete not found: {target_path}")
 
             elif self.button_input == 'SEARCH_BACKUP':
                 _search_start_sb = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _search_start_sb = datetime.now()
                     print(f"DEBUG: execute SEARCH_BACKUP START")
                 # Path to the directory containing Blender version folders (e.g., .../Blender/3.6, .../Blender/4.0)
@@ -1737,50 +933,50 @@ class OT_BackupManager(Operator):
 
                 pref_backup_versions.clear()
                 _fv1_start_sb = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _fv1_start_sb = datetime.now()
                     print(f"DEBUG: execute SEARCH_BACKUP calling find_versions for blender_versions_parent_dir: {blender_versions_parent_dir}")
-                found_backup_versions = find_versions(blender_versions_parent_dir)
-                if prefs().debug:
+                found_backup_versions = utils.find_versions(blender_versions_parent_dir) # Use utils
+                if pref_instance.debug:
                     _fv1_end_sb = datetime.now()
                     print(f"DEBUG: (took: {(_fv1_end_sb - _fv1_start_sb).total_seconds():.6f}s) execute SEARCH_BACKUP find_versions for blender_versions_parent_dir DONE")
                 pref_backup_versions.extend(found_backup_versions)
                 pref_backup_versions.sort(reverse=True)
 
                 # For restore_versions, search within the system_id folder in the backup_path
-                system_specific_backup_path = os.path.join(prefs().backup_path, prefs().system_id)
+                system_specific_backup_path = os.path.join(pref_instance.backup_path, pref_instance.system_id)
                 pref_restore_versions.clear()
                 _fv2_start_sb = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _fv2_start_sb = datetime.now()
                     print(f"DEBUG: execute SEARCH_BACKUP calling find_versions for backup_path: {system_specific_backup_path}")
-                combined_restore_versions = find_versions(system_specific_backup_path) + pref_backup_versions
-                if prefs().debug:
+                combined_restore_versions = utils.find_versions(system_specific_backup_path) + pref_backup_versions # Use utils
+                if pref_instance.debug:
                     _fv2_end_sb = datetime.now()
                     print(f"DEBUG: (took: {(_fv2_end_sb - _fv2_start_sb).total_seconds():.6f}s) execute SEARCH_BACKUP find_versions for backup_path DONE")
                 # Use dict.fromkeys to preserve order of first appearance if that's desired before sorting
                 pref_restore_versions.extend(list(dict.fromkeys(combined_restore_versions)))
                 pref_restore_versions.sort(reverse=True)
-                if prefs().debug and _search_start_sb:
+                if pref_instance.debug and _search_start_sb:
                     _search_end_sb = datetime.now()
                     print(f"DEBUG: (took: {(_search_end_sb - _search_start_sb).total_seconds():.6f}s) execute SEARCH_BACKUP END")
 
             elif self.button_input == 'SEARCH_RESTORE': 
                 _search_start_sr = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _search_start_sr = datetime.now()
                     print(f"DEBUG: execute SEARCH_RESTORE START")
                 blender_versions_parent_dir = os.path.dirname(bpy.utils.resource_path(type='USER'))
 
                 # For restore_versions, search within the system_id folder in the backup_path
-                system_specific_backup_path = os.path.join(prefs().backup_path, prefs().system_id)
+                system_specific_backup_path = os.path.join(pref_instance.backup_path, pref_instance.system_id)
                 pref_restore_versions.clear()
                 _fv1_start_sr = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _fv1_start_sr = datetime.now()
                     print(f"DEBUG: execute SEARCH_RESTORE calling find_versions for backup_path: {system_specific_backup_path}")
-                found_restore_versions = find_versions(system_specific_backup_path)
-                if prefs().debug:
+                found_restore_versions = utils.find_versions(system_specific_backup_path) # Use utils
+                if pref_instance.debug:
                     _fv1_end_sr = datetime.now()
                     print(f"DEBUG: (took: {(_fv1_end_sr - _fv1_start_sr).total_seconds():.6f}s) execute SEARCH_RESTORE find_versions for backup_path DONE")
                 pref_restore_versions.extend(found_restore_versions)
@@ -1788,15 +984,15 @@ class OT_BackupManager(Operator):
 
                 pref_backup_versions.clear()
                 _fv2_start_sr = None
-                if prefs().debug:
+                if pref_instance.debug:
                     _fv2_start_sr = datetime.now()
                     print(f"DEBUG: execute SEARCH_RESTORE calling find_versions for blender_versions_parent_dir: {blender_versions_parent_dir}")
-                combined_backup_versions = find_versions(blender_versions_parent_dir) + pref_restore_versions
-                if prefs().debug:
+                combined_backup_versions = utils.find_versions(blender_versions_parent_dir) + pref_restore_versions # Use utils
+                if pref_instance.debug:
                     _fv2_end_sr = datetime.now()
                     print(f"DEBUG: (took: {(_fv2_end_sr - _fv2_start_sr).total_seconds():.6f}s) execute SEARCH_RESTORE find_versions for blender_versions_parent_dir DONE")
 
-                if prefs().debug:
+                if pref_instance.debug:
                     print("Combined backup versions before filtering: ", combined_backup_versions)
                 
                 # Filter and sort backup versions
@@ -1807,14 +1003,14 @@ class OT_BackupManager(Operator):
                         float(version_tuple[0]) # Check if version name can be a float
                         valid_backup_versions.append(version_tuple)
                     except ValueError:
-                        if prefs().debug:
+                        if pref_instance.debug:
                             print(f"Filtered out non-float-like version from backup_versions: {version_tuple[0]}")
                 
                 pref_backup_versions.extend(valid_backup_versions)
-                if prefs().debug:
+                if pref_instance.debug:
                     print("Final backup_versions list: ", pref_backup_versions)
                 pref_backup_versions.sort(reverse=True)
-                if prefs().debug and _search_start_sr:
+                if pref_instance.debug and _search_start_sr:
                     _search_end_sr = datetime.now()
                     print(f"DEBUG: (took: {(_search_end_sr - _search_start_sr).total_seconds():.6f}s) execute SEARCH_RESTORE END")
